Refactored: add CommentList
