+++ src/EditBox.js
@@ -3,8 +3,6 @@ import React, {useState} from "react";
 export default function EditBox({text}){
     const [edit, setEdit] = useState(text)
     return (
-//         <>
-//   <label>Edit
   <input
       id='drinkName'
       size={edit.length}
@@ -12,7 +10,5 @@ export default function EditBox({text}){
       value={edit}
       onChange={(e) => setEdit(e.target.value)}
   />
-//   </label>
-//   </>
 )
 }
